FPV-23 code clean up and modularizing
