--- web-intf/backend/hooks.js
@@ -1,11 +1,0 @@
-const bcrypt = require("bcrypt")
-
-function compareHashPassword(hashedPassword, password){
-
-    bcrypt.compare(password,hashedPassword).then((result) => {
-        console.log(result)
-        return result;
-    })
-}
-
-module.exports = { compareHashPassword };
