+++ web-intf/backend/routes/routes.js
@@ -2,42 +2,13 @@ const express = require("express");
 const router = express.Router();
 const User = require('../models/User');
 const JWT = require('../models/JWT');
-const {checkInactiveToken, checkUserValidations} = require('../utilities/utilities');
-const bcrypt = require("bcrypt");
 const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv');
-
+const {checkJwtExpiration, checkUserValidations} = require('../utilities/utilities');
+const bcrypt = require("bcrypt");
 
 dotenv.config();
 const JWT_KEY = process.env.JWT_SECRET;
-
-const checkJwtExpiration = async (req, res, next) => {
-    const token = req.headers.authorization?.split(' ')[1];
-    if (token) {
-        try {
-            const decodedToken = jwt.verify(token, JWT_KEY);
-            const matchToken = await checkInactiveToken(token);
-
-            if (matchToken === true) {
-                return res.status(401).json({ message: 'Token is inactive' });
-            }
-            else if (decodedToken.exp * 1000 < Date.now()) {
-                // Token has expired, send a 401 Unauthorized response
-                return res.status(401).json({message: 'Token has expired'});
-            }
-
-            // Token is still valid, continue with the request
-            req.user = decodedToken;
-            next();
-        } catch (err) {
-            return res.status(401).json({ message: 'Invalid token' });
-        }
-    } else {
-        return res.status(401).json({ message: 'Token not provided' });
-    }
-};
-
-
 
 router.post('/register', async (req, res) => {
     const username = req.body.username;
@@ -61,11 +32,9 @@ router.post('/register', async (req, res) => {
         return res.status(400).send({ error: "User already exists!" });
 
     } catch (error) {
-        console.log(error);
         return res.status(500).send({ error: "Unable To Create User!" });
     }
 })
-
 
 router.post('/login', async (req, res) => {
     const username = req.body.username;
@@ -79,7 +48,7 @@ router.post('/login', async (req, res) => {
             const match = await bcrypt.compare(password, user[0].password);
             if (match) {
                 const token = jwt.sign({ user: username }, JWT_KEY, { expiresIn: '1hr' });
-                return res.status(200).send({ token: token });
+                return res.status(200).send({ token: token, userId: user[0].id });
             } 
                 return res.status(401).send({ error: "Incorrect Password!" });
         }
@@ -105,8 +74,6 @@ router.post('/logout', checkJwtExpiration, async (req, res) => {
     return res.sendStatus(401);
 
 });
-
-
 
 //Put route to change user password
 module.exports = router;
